Moved controller and removed unused dep.
