+++ pom.xml
@@ -94,11 +94,6 @@
             <scope>test</scope>
         </dependency>
         <dependency>
-            <groupId>io.projectreactor</groupId>
-            <artifactId>reactor-test</artifactId>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
             <groupId>com.two</groupId>
             <artifactId>testbed</artifactId>
             <version>${testbed.version}</version>
